+++ internal/htmlcheck/inline_test.go
@@ -72,7 +72,7 @@ func TestInlineStyleDetection(t *testing.T) {
 	}
 }
 
-func assertEqual(t *testing.T, a interface{}, b interface{}, message string) {
+func assertEqual(t *testing.T, a any, b any, message string) {
 	if a == b {
 		return
 	}